apps/worker/pyproject.toml
[project]
name = "expense-budget-tracker-worker"
version = "0.1.0"
description = "Exchange rate fetcher worker for expense-budget-tracker"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary>=2.9",
    "requests>=2.31",
    "schedule>=1.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "mypy>=1.10",
    "types-requests>=2.31",
]
